display customer logo in virto header
