--- src/VirtoAppHeader/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./VirtoAppHeader";
-export type { VirtoAppHeaderProps } from "./VirtoAppHeader.types";
